changed result component to Class and added the drink to be displayed to the search by name case, and added models to be displayed when img is clicked on
