+++ src/App.js
@@ -42,16 +42,16 @@ class App extends Component {
                   title: result.strDrink,
                   img: result.strDrinkThumb,
                   instructions: result.strInstructions,
-                  ingredient1: result.strIngredient1,
-                  ingredient2: result.strIngredient2,
-                  ingredient3: result.strIngredient3,
-                  ingredient4: result.strIngredient4,
-                  ingredient5: result.strIngredient5,
-                  ingredient6: result.strIngredient6,
-                  ingredient7: result.strIngredient7,
-                  ingredient8: result.strIngredient8,
-                  ingredient9: result.strIngredient9,
-                  ingredient10: result.strIngredient10,
+                  ingredientsArr: [result.strIngredient1, result.strIngredient2, result.strIngredient3, result.strIngredient4, result.strIngredient5,],
+                  // ingredient2: result.strIngredient2,
+                  // ingredient3: result.strIngredient3,
+                  // ingredient4: result.strIngredient4,
+                  // ingredient5: result.strIngredient5,
+                  // ingredient6: result.strIngredient6,
+                  // ingredient7: result.strIngredient7,
+                  // ingredient8: result.strIngredient8,
+                  // ingredient9: result.strIngredient9,
+                  // ingredient10: result.strIngredient10,
                   measure1: result.strMeasure1,
                   measure2: result.strMeasure2,
                   measure3: result.strMeasure3,
@@ -119,16 +119,16 @@ class App extends Component {
                   title: result.strDrink,
                   img: result.strDrinkThumb,
                   instructions: result.strInstructions,
-                  ingredient1: result.strIngredient1,
-                  ingredient2: result.strIngredient2,
-                  ingredient3: result.strIngredient3,
-                  ingredient4: result.strIngredient4,
-                  ingredient5: result.strIngredient5,
-                  ingredient6: result.strIngredient6,
-                  ingredient7: result.strIngredient7,
-                  ingredient8: result.strIngredient8,
-                  ingredient9: result.strIngredient9,
-                  ingredient10: result.strIngredient10,
+                  ingredientsArr: [result.strIngredient1, result.strIngredient2, result.strIngredient3],
+                  // ingredient2: result.strIngredient2,
+                  // ingredient3: result.strIngredient3,
+                  // ingredient4: result.strIngredient4,
+                  // ingredient5: result.strIngredient5,
+                  // ingredient6: result.strIngredient6,
+                  // ingredient7: result.strIngredient7,
+                  // ingredient8: result.strIngredient8,
+                  // ingredient9: result.strIngredient9,
+                  // ingredient10: result.strIngredient10,
                   measure1: result.strMeasure1,
                   measure2: result.strMeasure2,
                   measure3: result.strMeasure3,
